// config/enums/Mountains.ts
export enum Mountain {
	CrystalMountain = 'Crystal Mountain',
	SummitAtSnoqualmie = 'Summit at Snoqualmie',
	MtBaker = 'Mt Baker',
	StevensPass = 'Stevens Pass',
	MtHood = 'Mt Hood',
	MtBachelor = 'Mt Bachelor',
	SunValley = 'Sun Valley',
	BigSky = 'Big Sky',
	Vail = 'Vail',
	MammothMountain = 'Mammoth Mountain',
	ParkCity = 'Park City',
	ArapahoeBasin = 'Arapahoe Basin',
	Breckenridge = 'Breckenridge',
	CopperMountain = 'Copper Mountain',
	Keystone = 'Keystone',
	Loveland = 'Loveland',
	WinterPark = 'Winter Park',
	Steamboat = 'Steamboat',
	AspenSnowmass = 'Aspen Snowmass',
	Telluride = 'Telluride',
	CrestedButte = 'Crested Butte',
	Purgatory = 'Purgatory',
	WolfCreek = 'Wolf Creek',
	Silverton = 'Silverton',
	Powderhorn = 'Powderhorn',
}

export enum States {
	Washington = 'Washington',
	Oregon = 'Oregon',
	Idaho = 'Idaho',
	Montana = 'Montana',
	Colorado = 'Colorado',
	California = 'California',
	Utah = 'Utah',
}